Fix D6 roll count display and draw Compact SeedQR at its true size
The dice-roll screen printed "zu rolls needed" instead of the count: nano
newlib formatting (kept for the 285 KiB image budget) doesn't understand
%zu, so the specifier leaked through literally.

Compact SeedQR always rendered at the fixed QR version sized for the
account-key/xpub screen instead of the smallest version its own entropy
needs, defeating the "compact" of the SeedSigner/Krux convention it
follows. Unlocking the encoder's version table costs under 2 KiB of
flash, still comfortably inside the audited budget. The draw scale and
title placement are now derived from the QR actually drawn rather than a
size baked in for the larger code, so a smaller code fills the display
instead of leaving a dead gap next to a squeezed title.

+++ main/seedtool_app.c
@@ -1030,7 +1030,7 @@ static int collect_entropy(const int source, const size_t words)
         /* Shown once, with the bar already in place but empty, so its shape and
          * position are seen before they start moving. */
         char rolls_needed[24];
-        (void)snprintf(rolls_needed, sizeof(rolls_needed), "%zu rolls needed", required);
+        (void)snprintf(rolls_needed, sizeof(rolls_needed), "%u rolls needed", (unsigned)required);
         const seedtool_progress_t empty = { 0 };
         if (!dice_confirm(names[source], rolls_needed, "Red bar = non-random", &empty)) {
             outcome = 0;
